% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeBase.R
\docType{data}
\name{MSigDB_GS}
\alias{MSigDB_GS}
\title{Canonical pathways defined in MSigDB}
\format{check vignettes for details}
\usage{
data(MSigDB_GS)
}
\description{
Canonical pathways defined in MSigDB
}
\keyword{datasets}
