% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnowledgeBase.R
\docType{data}
\name{MSigDB_NetORA_GS}
\alias{MSigDB_NetORA_GS}
\title{Network neighbors of canonical pathways defined in MSigDB (FDR 0.05)}
\format{check vignettes for details}
\usage{
data(MSigDB_NetORA_GS)
}
\description{
Network neighbors of canonical pathways defined in MSigDB (FDR 0.05)
}
\keyword{datasets}
