% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetORA_MSigDB_CP.R
\name{NetORA_MSigDB_CP}
\alias{NetORA_MSigDB_CP}
\title{Network-based enrichment for a list of signatures}
\usage{
NetORA_MSigDB_CP(SG, MSigDB_NetORA_GS, PG)
}
\arguments{
\item{SG}{a vector of gene signature}

\item{MSigDB_NetORA_GS}{pre-computed neighbors of Canonical pathways as defined in MSigDB}

\item{PG}{a vector of background genes}
}
\value{
a list
}
\description{
This function is wraper of core function NetORA to work on a list of gene sets
}
\examples{
library("NetORA")
data(text_2006_12_NetGPA)
data(Example_NetGPA)
data(MSigDB_NetORA_GS)
ExE_Hyper  <- Example_NetGPA$ExE_Hyper
PG <- colnames(text_2006_12_NetGPA)
CancerPathway <- paste0("REACTOME_", Example_NetGPA$CancerPathway)
mergedT <- NetORA_MSigDB_CP(ExE_Hyper, MSigDB_NetORA_GS[CancerPathway], PG)
}
\references{
Jiantao Shi: NetORA, a package for network-based pathway over-representation analysis.
}
\author{
Jiantao Shi
}
