% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetORA_pHyper.R
\name{NetORA_pHyper}
\alias{NetORA_pHyper}
\title{The Hypergeometric Distribution in NetORA}
\usage{
NetORA_pHyper(SG, SR, PG)
}
\arguments{
\item{SG}{a vector of signature}

\item{SR}{a vector of pathway}

\item{PG}{a vector of background genes}
}
\value{
a list
}
\description{
This function implements a test based on Hypergeometric Distribution
}
\examples{
library("NetORA")
data(text_2006_12_NetGPA)
data(Example_NetGPA)
ExE_Hyper  <- Example_NetGPA$ExE_Hyper
queryTable <- NetORA_Pre(ExE_Hyper, text_2006_12_NetGPA, progressBar = TRUE)
Cancer_GS  <- Example_NetGPA$Cancer_GeneSet
PG <- colnames(text_2006_12_NetGPA)
mergedT <- NetORA_GS(Cancer_GS, queryTable, PG, FDR = 0.05)
}
\references{
Jiantao Shi: NetORA, a package for network-based pathway over-representation analysis.
}
\author{
Jiantao Shi
}
