% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetORA_Pre.R
\name{NetORA_Pre}
\alias{NetORA_Pre}
\title{Building sub-network in background}
\usage{
NetORA_Pre(SR, netMatrix, progressBar = TRUE)
}
\arguments{
\item{SR}{a vector of pathway}

\item{netMatrix}{a integer matrix which is accepted by NetGPA as a network}
}
\value{
a data frame
}
\description{
This function implements a preprocessing step that find all neighbor of a given pathway.
}
\examples{
library("NetGPA")
data(text_2006_12_NetGPA)
data(Example_NetGPA)
ExE_Hyper  <- Example_NetGPA$ExE_Hyper
queryTable <- NetORA_Pre(ExE_Hyper, text_2006_12_NetGPA, progressBar = TRUE)
Cancer_GS  <- Example_NetGPA$Cancer_GeneSet
PG <- colnames(text_2006_12_NetGPA)
mergedT <- NetORA_GS(Cancer_GS, queryTable, PG, FDR = 0.05)
}
\references{
Jiantao Shi: NetORA, a package for network-based pathway over-representation analysis.
}
\author{
Jiantao Shi
}
